Improve secret handling
The isManagedTLSSecret helper relied on ingresses, but wasn't checked on
an update to ingresses. This instead moves all checks to the plugin.

In addition, creating a secret informer builds a cache of all secrets
in memory. This is undesirable, because it means we keep around
sensitive data from completely unrelated secrets.

Instead, a transform is installed on the informer to throw away all
secret data, so we only keep metadata. The plugin is then responsible
for fetching actual secret data for only the secrets we need.

In practice, we are still dependant on the Go garbage collector.

+++ internal/controller/queued_event_handler.go
@@ -21,14 +21,21 @@ type eventAction[T any] struct {
 func (e eventAction[T]) handle(c *CaddyController) error {
 	switch e.action {
 	case actionAdd:
-		return e.handlers.AddFunc(e.resource)
+		if e.handlers.AddFunc != nil {
+			return e.handlers.AddFunc(e.resource)
+		}
 	case actionUpdate:
-		return e.handlers.UpdateFunc(e.resource)
+		if e.handlers.UpdateFunc != nil {
+			return e.handlers.UpdateFunc(e.resource)
+		}
 	case actionDelete:
-		return e.handlers.DeleteFunc(e.resource)
+		if e.handlers.DeleteFunc != nil {
+			return e.handlers.DeleteFunc(e.resource)
+		}
 	default:
 		panic("invalid action")
 	}
+	return nil
 }
 
 type QueuedEventHandlers[T any] struct {
